Set up a basic set of unit target test functions.
Since a lot of target menus are going to be looking for units with
a particular team allegiance, I decided to set up a class full of static
functions that can be used by anything that needs one of these
basic unit targeting tests.

This commit also sets up the TargetTypes enum, which may be used
in the future to help classify objects by what type of unit targeting they
might use. I'm imagining it being pretty useful for items in particular,
so it is easy to distinguish between items that can be used on allies,
enemies, anyone, etc.