Created AttackTargetMenu class.
This commit implements the first of many target menus.
This one is pretty simple right now; it just overrides the
logic of the template's refreshTargets function so that
it can get the data it needs for its possibleTargets array.

This takes advantage of functions and infrastructure set up in
previous commits, with other misc. changes needed to make
this work included in here. The changes should, on the whole,
be pretty straightforward to follow. If they aren't, don't worry, I'll
give them proper documentation at some future point in time.