Replaced the old, fake attack target menu with the new class.
Previously, the menuManager's "AttackTargetMenu" was just a ResizableBasicMenu
with a few static "targeting" options to pick from. With these changes, the menuManager
actually uses the AttackTargetMenu properly.

To make the menus properly "refresh" their set of options upon opening, I set up the
openFunctions array, similar to the confirmFunctions and cancelFunctions I put together
in the past. In those functions, each menu can do whatever checks are necessary to
set up their appearance/options to match whatever the current situation is.

At the moment, open functions are only implemented for a basic version of attack targeting,
but this solution can be extended to work for all of the menus under the MenuManager.