Added a info display to AttackTargetMenu.
This uses the data from the MissionState's selected unit and
the target unit to display the combat statistics of both characters.

Overall, I think this sort of strucutre is what I'll be using for the other target
menus too. If there's some functionality that's shared by all unit-targeting
(or all targeting) menus, I'll make sure to migrate that functionality into a
common class.