Made the AttackTargetMenu use the selected unit's equipped weapon upon open.
In the process, I also ironed out a bit of an issue that had come up in earlier code.
I needed to set the weapon's information, but the weapon information setup involved
looking information about the selected and target units, which weren't set up yet.

I got out of that catch-22 by adding an if to the weapon info-setting code, so it
will only look up unit information if the unit properties of the menu aren't null.