Integrated InvBoxes into the TradeMenu.
Previously, the TradeMenu used its own logic to set up two connected
inventory windows, which was totally separate from the logic used to
create the inventory shown when an "InventoryMenu" is created.

Now both the InventoryMenu and TradeMenu use a single class,
InventoryBox, to handle the logic of setting up an inventory window.
This makes it easier to change up the look of our inventory windows
in the future, since there will only be one point of entry that needs changing.