Added a function to update equippable index arrays in the inventory.
Right now it's just being used after trading to update all index information,
which is the only time I need to look over the whole inventory to fix it right now.

Eventually I'll have to handle the case where a single item can be dropped out
of an inventory, but luckily that's a ways off from being an issue.