Added system for updating a unit's attack ranges.
When the inventory observes a change in a unit's items, it flips a boolean
value in its owning unit to signal that a change has (potentially) occurred.

When any other class attempts to access a unit's attack/heal range
(now through a getter function instead of direct access) the unit will check
if it needs to re-identify its attack and heal ranges, and does so if necessary.

It's important to remember that future classes MUST correctly set the unit's
itemsHaveChanged boolean value to true if any range-altering change may
have happened to the unit's inventory, otherwise inccorrect attack/heal ranges
may be displayed.