Fixed bug with vertical menu option wrapping.
Previously, menu options were not setting their "downIsWrap"
variable back to false if menuOptions below them became
valid again. This meant that holding down wouldn't always
let the cursor move to the bottom of the menu before stopping.

This has now been fixed for both the InventoryBox and
ResizableBasicMenu, which were the two places where
this logic was explicitly described.