Fixed TradeMenu post-trade & cancel cleanup.
Previously, the player could select an item, move the
cursor over one of the dummy items, and press the
"cancel" input and end up with the cursor hovering over
a no-longer visible MenuOption.

This fixes it by jumping the cursor back to the selected item
when the player cancels out of a trade. It also does some
cleanup with nulling out variables, reducing redundant logic
in post-trade and cancel functions.