Completed basic combat menu implementation.
This handles beginning, continuing, and ending combat
appropriately. All computations are done based on FE7-
style combat formulae (with some missing features)
and will probably not be final.

Timing and aesthetic-related values are likely to change
in the next day or so of commits, but the logical skeleton
should be (for the most part) unchanged.