Fixed bug with AttackTargetMenu's second name box position/sizing.
Issue was caused by name boxs' logical positions not getting updated properly to match the sprites
after setPos() was called for the whole menu.

This really exposes the problem with my menu class hierarchy. Some more design work and refactoring
needs to go into this stuff so that the menu repositioning logic is more intuitive for menus that contain
more than one resizable and/or moveable box.